Do not install libkeys.a
libkeys.a is a static library therefore it shouldn't be installed
in the system as part of the regular "install" rules. There should
be a "install-dev" (or so) to install this library together with
the header files required for development.
For now, we will just remove that installation to remove the
lintian warning.

Signed-off-by: Mathieu Olivari <mathieu.olivari@gmail.com>