Change default manpage installation destination to /usr/local hierarchy
Since we install the binary programs to /usr/local hierarchy by default,
we should also use the /usr/local hierarchy for manpages and do not mess
with /usr/share/man directories which are usually managed by distributions.

This allows better tracking of user installed manpages and is a cleaner
approach in general.

Signed-off-by: Michael Heimpold <mhei@heimpold.de>