slac: allow to focus on sounding only
This enhances the evse and pev tool to skip the step which sets
the NMK and other stuff and focus on the sounding part only to
determine attenuation. This can be used for example during a
device test bed, where only the sounding step should be run to
speedup the whole process.

Signed-off-by: Michael Heimpold <michael.heimpold@i2se.com>