plcrate: introduce command line parameter -s
This introduces the commmand line switch -s which allows to
fine-control the network traffic generation: it suppresses
the generation of "antiphon" traffic, i.e. RX traffic (from
remote peer to local one) and thus speed-ups network testing
scenarios. This flag is only used in combination with
-t flag (one-to-many case) - the many-to-many case (-T) is
not affected.

For this is (re-)uses a flag PLC_TXONLY which requires minor
changes in amprate to keep existing behaviour of amprate.
(this flag was not usable in amprate - so I guess it is ok).

Also the man page is updated for plcrate to document this new
parameter.

Signed-off-by: Michael Heimpold <michael.heimpold@i2se.com>