plcstat/int6kstat: allow to query a powerline device for topology directly
In a switched network, i.e. when local powerline device(s) are not
connected directly to host, then using the Local Management Address
to determine the local devices might fail due to wrong or incomplete
forwarding by the switch.

In this case it would be desirable to address the local powerline
device directly which is not possible with both tools at the moment.

This patch fixes this issue by keeping current behaviour as much as
possible and also document it in the corresponding man pages.

Signed-off-by: Michael Heimpold <michael.heimpold@i2se.com>