Fix parsing of command line options for {amp,int6k,plc}rate and int6kbaud
When compiling with -Wimplicit-fallthrough= I noticed that compiler
warned about several statements which may fall through. After inspection
I'm sure that falling through is not desired behavior but simply
a break is missing in all cases.

So let's fix the unwanted side-effects which occur at the moment
when using the affected command line options.

Signed-off-by: Michael Heimpold <mhei@heimpold.de>